radius

+++ src/Engine/WorkingDir/shaders.glsl
@@ -49,7 +49,7 @@ struct Light
 	vec3 direction;
 	float intensity;
 	vec3 position;
-	int radius;
+	float radius;
 };
 
 layout(binding = 0, std140) uniform GlobalParams
@@ -94,7 +94,7 @@ struct Light
 	vec3 direction;
 	float intensity;
 	vec3 position;
-	int radius;
+	float radius;
 };
 
 layout(binding = 0, std140) uniform GlobalParams
